+++ Pods/Bugsnag/Source/Bugsnag/BugsnagCrashReport.h
@@ -10,12 +10,32 @@
 
 @interface BugsnagCrashReport : NSObject
 
-@property(readonly) NSDictionary *ksReport;
+- (id)initWithKSReport:(NSDictionary*)report;
 
-- (id)initWithKSReport:(NSDictionary *)report;
-- (NSDictionary *)serializableValueWithTopLevelData:(NSMutableDictionary *)data;
+@property NSDictionary *ksReport;
 
-- (NSString *)releaseStage;
-- (NSArray *)notifyReleaseStages;
+@property (readonly) NSString *releaseStage;
+@property (readonly) NSArray *notifyReleaseStages;
+@property (readonly) NSString *context;
+@property (readonly) NSString *appVersion;
+
+@property (readonly) NSArray *binaryImages;
+@property (readonly) NSArray *threads;
+@property (readonly) NSArray *breadcrumbs;
+@property (readonly) NSDictionary *error;
+@property (readonly) NSString *errorType;
+@property (readonly) NSString *errorClass;
+@property (readonly) NSString *errorMessage;
+@property (readonly) NSString *severity;
+@property (readonly) NSString *dsymUUID;
+@property (readonly) NSString *deviceAppHash;
+@property (readonly) NSUInteger depth;
+
+@property (readonly) NSDictionary *metaData;
+@property (readonly) NSDictionary *appStats;
+
+// These shouldnt be exposed
+@property (readonly) NSDictionary *system;
+@property (readonly) NSDictionary *state;
 
 @end